requirements.txt
Django==4.1.7
django-tailwind
django-crispy-forms
crispy-tailwind
django-address
django-cities-light
django-watson
django-autocomplete-light
django-filter
django-tables2